requirements.txt
matplotlib==2.2.3
numpy==1.18.5
pandas==0.23.4
Pillow==9.0.1
torch==1.9.0
torchvision==0.8.2
tqdm==4.26.0
